don't tally anatsite testing/pos for CDC scenario
The prob.xGC.tested and prop.%anatsite%.tested parameters look for
ids_seekTest, an object that does not exist in the "cdc" STI screening
scenario. So we wrap calculations at the end of the module which rely on
ids_seekTest in an if statement that checks to make sure "cdc" is not
the selected protocol.